+++ libs/ui/src/lib/button/button.spec.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
+import { RouterModule } from '@angular/router';
 import { NsButtonComponent } from './button';
 
 @Component({
@@ -15,7 +16,7 @@ describe('NsButtonComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [NsButtonComponent],
+      imports: [NsButtonComponent, RouterModule.forRoot([])],
     }).compileComponents();
 
     fixture = TestBed.createComponent(NsButtonComponent);